+++ music21/midi/realtime.py
@@ -18,7 +18,7 @@ https://joecodeswell.wordpress.com/2012/06/13/how-to-produce-python-controlled-a
 
 https://stackoverflow.com/questions/10983462/how-can-i-produce-real-time-audio-output-from-music-made-with-music21
 
-Requires pygame: http://www.pygame.org/download.shtml
+Requires pygame: pip3 install pygame
 '''
 from __future__ import annotations
